- include an index.html in the archive directory that, by default,       simply explains briefly what the archive is and disclaims responsability       for the content in it.  archive operators can customize the file as they       see fit by editing "$rootDir/archive/index.html".  That index will of       course be visible to anyone viewing the archive on the web, and it is       also published inside freenet archives (as are any other html files)     - more aggressively clean up temporary files on syndie shutdown     - fixed a regression where fetched passphrase encrypted posts did not       properly prompt the user for the passphrase     - add support for HTTP 301 redirects to EepGet (so freenet USKs work as       intended)

+++ src/syndie/db/HTTPSyndicator.java
@@ -246,7 +246,7 @@ public class HTTPSyndicator {
             try {
                 NestedUI nested = new NestedUI(_ui);
                 ok = imp.processMessage(nested, new FileInputStream(f), _client.getLoggedInNymId(), _client.getPass(), null);
-                if (ok && (nested.getExitCode() >= 0)) {
+                if (ok && (nested.getExitCode() >= 0) && (!imp.wasPBE()) ) {
                     _ui.debugMessage("Import successful for " + uri);
                     f.delete();
                     imported++;
@@ -276,12 +276,12 @@ public class HTTPSyndicator {
     public SyndieURI getMissingURI(int index) { return (SyndieURI)_pendingPBEURIs.get(index); }
     public void importPBE(int index, String passphrase) {
         Importer imp = new Importer(_client, _client.getPass());
-        File f = (File)_pendingPBEFiles.get(index);
         if ( (_pendingPBEURIs.size() <= index) || (index < 0) ) {
             _ui.errorMessage("Index out of bounds");
             _ui.commandComplete(-1, null);
             return;
         }
+        File f = (File)_pendingPBEFiles.get(index);
         SyndieURI uri = (SyndieURI)_pendingPBEURIs.get(index);
         boolean ok;
         try {
